fix(matches): share match form component
